Refactor build generator out
